+++ ndk-cache-redis/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @nostr-dev-kit/ndk-cache-redis
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+    - @nostr-dev-kit/ndk@2.14.4
+
 ## 0.1.2
 
 ### Patch Changes