+++ ndk-svelte/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @nostr-dev-kit/ndk-svelte
 
+## 2.4.10
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+    - @nostr-dev-kit/ndk@2.14.4
+
 ## 2.4.9
 
 ### Patch Changes